Update bindgen to v0.58
bindgen v0.54 transitively pulls in syntex_syntax v0.58.1, which no
longer compiles on Rust v1.41+:

https://github.com/rust-lang/rust/issues/68729

I'm hoping to use libpurple-rust in a rust-native reimplementation of
https://code.ur.gs/lupine/purple-plugin-delta , which will need to
build against Rust v1.50, so fixing this would be nice.

I fixed a few deprecation warnings in build.rs, but otherwise this
seems to be a simple update.